[Bronze V] Title: 이상한 기호, Time: 52 ms, Memory: 5488 KB, Score: 100 point -BaekjoonHub

+++ 백준/Bronze/15964. 이상한 기호/이상한 기호.cs
@@ -6,7 +6,7 @@ namespace baejoon_15964
     {        
         static void Main()
         {
-            int[] iArr = Array.ConvertAll(Console.ReadLine().Split(), int.Parse);
+            long[] iArr = Array.ConvertAll(Console.ReadLine().Split(), long.Parse);
             Console.WriteLine((iArr[0] + iArr[1]) * (iArr[0] - iArr[1]));
         }
     }
